# doc/api/openapi.yaml
openapi: 3.0.3
info:
  title: Blog System API
  description: Blog Management System API Documentation
  version: 1.0.0
tags:
  - name: Auth
    description: Authentication and Authorization API
  - name: User
    description: User Management API
  - name: Article
    description: Article Management API
  - name: Category
    description: Category Management API
  - name: Tag
    description: Tag Management API
  - name: Comment
    description: Comment Management API
  - name: Interact
    description: Interaction Management API
  - name: Favorite
    description: Favorite Management API
  - name: Config
    description: Configuration Management API
paths:
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: User login to get access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultLoginVO'
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: User logout
      description: User logout to invalidate token
      operationId: logout
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh token
      description: Refresh access token with current token
      operationId: refreshToken
      parameters:
        - name: Authorization
          in: header
          description: Authorization token
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
components:
  schemas:
    LoginReq:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    LoginVO:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserVO'
    UserVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        nickname:
          type: string
        email:
          type: string
        avatar:
          type: string
    ResultLoginVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/LoginVO'
    ResultVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ResultString:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: string